+++ src/main/java/at/hannibal2/skyhanni/SkyHanniMod.java
@@ -231,6 +231,7 @@ public class SkyHanniMod {
         loadModule(new WrongFungiCutterWarning());
         loadModule(new FarmingArmorDrops());
         loadModule(new JoinCrystalHollows());
+        loadModule(new GardenVisitorColorNames());
 
         Commands.INSTANCE.init();